fixes36.1 update: #H4725 - repeated shortcut keys
Mark an old 'unconfirmed' bug as 'fixed'.  The report was for win32gui
and I can't reproduce the problem or verify that it's gone, but I'm
sure enough about the cause and long-ago fix to put #H4725 to rest....

Report was that the menu for the name-from-discoveries-list command
was reusing a set of punctuation characters for each class of objects.
The key was that '!' was always first, and it is (' ' + 1) so the core
bug of erroneously specifying space for a selector on object class
header/separator lines was being used as the start of the selector
sequence.  (Report that led to the fix was that typing space on that
menu always made it finish instead of advance to the next page.)